+++ index.html
@@ -10,7 +10,6 @@
 </head>
 
 <body>
-	<script src = 'js/main.js'></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -233,5 +232,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+ 
+  <script src = 'js/main.js' type = 'module'></script>
 </body>
 </html>
